MEDIUM: ssl: add alternative way to load certificates with io_uring
experimental patch to make use of IO_URING to batch load certificates;
this drastically reduces the number of syscall and might benefit to
setup with large number of certificates.
it uses liburing in order to batch operation as follow:
for each certificate directory, we apply the same operations on each
file:
 - statx
 - openat
 - read
 - close
the results are stored in an ebtree. Then when we need to load them
with the SSL lib, instead of providing a path, we provide a buffer to
be consumed. The tree is freed after each directory.

for now it requires a quite large limit of file descriptors, as all
operations types are done one after another; so the limit of fd should
be set higher than the number of certificates you have to load. This
part is probably going to evolve very soon as IO_URING plans are to be
able to chain operations with a given pre-defined file descriptor.

on a setup with 25k certificates I was able to measure a minimum of 20%
gain on the init time when the filesystem cache is empty. My testing is
as follow; for each directory:
- put a timer before `ssl_sock_load_cert_list_file`, including
  `ssl_load_certiodir` in the case of io_uring case.
- measure the time at the end of those operations, just after
  `ssl_free_certiodir` in the case of io_uring.

some results with my old ssd laptop, the average certificate init time
in ms, using 24686 certs:
▒▒▒▒▒▒▒▒▒▒▒▒▒▒           8554    io_uring_empty_cache
▒▒▒▒▒▒▒▒▒▒▒▒▒            8100    io_uring_full_cache
▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒      11395    syscall_empty_cache
▒▒▒▒▒▒▒▒▒▒▒▒▒            8087    syscall_full_cache

The benefits are of course less obvious with following reloads on a
machine where you only have haproxy process managing its own memory, but
it can be interesting on a busier setup where the filesystem cache is
often invalidated. However a very quick overview of reload time on our
production seems to show that more than 50% of the time, the files are
not present in cache anymore while reloading (I hope I'm not too far
from the reality in my analysis); so I would expect this patch could
participate to flatten our reload time; I however did not tested it in a
production environment.

It remains linux-specific though; the operations used require a minimum
of kernel >= v5.6 even though some operations were supported in v5.4, I
did not change the code to adapt the behavior. All my tests were
performed on >= v5.7rc4 as the time of writing.

it introduces three build parameters:
- USE_IO_URING
- IO_URING_INC
- IO_URING_LIB

Signed-off-by: William Dauchy <w.dauchy@criteo.com>